+++ js/src/forum/components/MinecraftLogInModal.js
@@ -1,16 +1,14 @@
 import Button from "flarum/common/components/Button";
 import Modal from "flarum/common/components/Modal";
-import Stream from "flarum/common/utils/Stream";
+import app from "flarum/forum/app";
 
-const trans = (key) => {
-    return app.translator.trans(`nearata-auth-minecraft.forum.${key}`);
+const trans = (key, options = {}) => {
+    return app.translator.trans(`nearata-auth-minecraft.forum.${key}`, options);
 };
 
 export default class MinecraftLogInModal extends Modal {
     oninit(vnode) {
         super.oninit(vnode);
-
-        this.token = Stream();
     }
 
     className() {
@@ -25,16 +23,14 @@ export default class MinecraftLogInModal extends Modal {
         return [
             m(".Modal-body", [
                 m(".Form.Form--centered", [
-                    m("p.helpText", trans("token_help")),
-                    m(".Form-group", [
-                        m("input.FormControl", {
-                            bidi: this.token,
-                            placeholder: trans("token_placeholder"),
-                            name: "token",
-                            type: "text",
-                            autocomplete: "off",
-                        }),
-                    ]),
+                    m(
+                        "p.helpText",
+                        trans("modal_help", {
+                            server: app.forum.attribute(
+                                "nearataMinecraftServerIp"
+                            ),
+                        })
+                    ),
                     m(".Form-group", [
                         m(
                             Button,
@@ -42,25 +38,11 @@ export default class MinecraftLogInModal extends Modal {
                                 type: "submit",
                                 className:
                                     "Button Button--primary Button--block",
-                                disabled: this.loading,
                                 loading: this.loading,
                             },
                             trans("log_in_modal.submit_button")
                         ),
                     ]),
-                ]),
-            ]),
-            m(".Modal-footer", [
-                m("span", [
-                    "Powered by ",
-                    m(
-                        "a",
-                        {
-                            href: "https://mc-oauth.net/",
-                            target: "_blank",
-                        },
-                        "Minecraft oAuth"
-                    ),
                 ]),
             ]),
         ];
@@ -74,22 +56,18 @@ export default class MinecraftLogInModal extends Modal {
         app.request({
             url: `${app.forum.attribute("apiUrl")}/auth/minecraft`,
             method: "POST",
-            body: { token: this.token() },
             errorHandler: this.onerror.bind(this),
         })
             .then((response) => {
                 app.authenticationComplete(response);
-                if ("loggedIn" in response) {
-                    window.location.reload;
-                }
             })
             .catch(() => {})
             .then(this.loaded.bind(this));
     }
 
     onerror(error) {
-        if (error.status === 403) {
-            error.alert.content = trans("error_403");
+        if (error.status === 400) {
+            error.alert.content = trans("login_error");
         }
 
         super.onerror(error);