+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- The user is no longer allowed to change the email.
+- From this version and up, this extension requires the Minecraft Plugin [OAuth](https://github.com/Neapovil/oauth) to allow user to authenticate.
+- The user no longer needs to type a TOTP to authenticate.
+
 ## [2.0.0] - 2021-06-20
 
 - Update to Flarum 1.0